Moved ConfigFile into the new Json namespace.
